refactor: enforce clippy -D warnings across all targets
Run the full lint workflow (fmt -> clippy -D warnings -> check -> test) and
fix every warning across 33 files (178 insertions, 168 deletions):

- matcher/mod.rs: honor MAX_ROOT_LOOKUPS (1633) in Phase 2 root enumeration
  (lookups counter was incremented but never read); &Vec -> &[_] params;
  doc-list formatting
- extractor/mod.rs: drop dead stage_skeleton/stage_skeleton_filters/
  stage_minutia_collection fns; disambiguate glob re-exports via aliases
  (absolute_contrast_compute, relative_contrast_compute, segmentation_compute)
- primitives: canonical Ord/PartialOrd for IntPoint (Ord is source of truth,
  PartialOrd delegates), Iterator::find for circular_list/reversed_list,
  needless_range_loop in circular_array::enlarge, clamp() in image_resize
- extractor/skeleton_filters_graph.rs: remove unused test helper long2_pts
- tests: replace manual min/max clamp in histogram_cube test with clamp()

AGENTS.md: add 5 lessons learned (lint workflow, clippy --fix + fmt,
partial_cmp canonical form, glob re-export aliasing, limit counter guards).

Verification: cargo fmt --check OK, clippy -D warnings 0 errors,
cargo check --all-targets clean, cargo test debug+release 463/463 passing.

+++ src/extractor/pixelwise_orientations.rs
@@ -134,7 +134,7 @@ pub fn compute(input: &DoubleMatrix, mask: &BooleanMatrix, blocks: &BlockMap) ->
                 let neighbor_set = &neighbors[(y as usize) % neighbors.len()];
                 for neighbor in neighbor_set {
                     let radius = neighbor.offset.x().abs().max(neighbor.offset.y().abs());
-                    let yr = y as i32;
+                    let yr = y;
                     if yr - radius >= 0 && yr + radius < input.height() as i32 {
                         let x_start = radius.max(valid_x_start);
                         let x_end = (input.width() as i32 - radius).min(valid_x_end);
